Refactor JournalEntry model to implement input/output methods
Formerly, the model took the entry as cents, and stored the amount in cents too.
Now the inputs formerly called cents1 and cents2 are called dollar1 and dollar2, and take in dollar values as inputs.
This way, the input corresponds better with the view, since we want the users to type the transaction amount in dollars, not cents.

This commit also implements some methods that are necessary to validate dollar inputs and
other methods that are needed to calculate the total debit/credit and amount for each transaction categories.